+++ src/core/World.lua
@@ -1,0 +1,6 @@
+--- src/core/World.lua
+-- @class World
+-- World and physics management
+local World = {}
+
+return World
