+++ src/entities/Enemy.lua
@@ -6,7 +6,7 @@ local ActiveEnemies = {}
 Enemy.__index = Enemy
 
 -- @import Player
-local Player = require("src.Player")
+local Player = require("src.entities.Player")
 
 -- Basically a getter method
 function Enemy:getCount()
