+++ src/entities/Stone.lua
