+++ src/visuals/Animation.lua
@@ -1,5 +1,6 @@
+--- src/visuals/Animation.lua
+-- @class Animation
+-- Animation system
 local Animation = {}
-
-
 
 return Animation
