+++ src/visuals/Background.lua
@@ -1,4 +1,6 @@
+--- src/visuals/Background.lua
+-- @class Background
+-- Background system with parallax effect
 local Background = {}
-
 
 return Background
